+++ packages/cli/src/commands/build-docs/utils.ts
@@ -1,4 +1,4 @@
-import { isAbsoluteUrl, logger, type Config } from '@redocly/openapi-core';
+import { isAbsoluteUrl, logger, type Config, HandledError } from '@redocly/openapi-core';
 import { default as handlebars } from 'handlebars';
 import { existsSync, lstatSync, readFileSync } from 'node:fs';
 import * as path from 'node:path';
@@ -7,7 +7,6 @@ import { renderToString } from 'react-dom/server';
 import { default as redoc } from 'redoc';
 import { ServerStyleSheet } from 'styled-components';
 
-import { exitWithError } from '../../utils/error.js';
 import { redocStandaloneSri } from '../../utils/package.js';
 import type { BuildDocsOptions } from './types.js';
 
@@ -52,11 +51,10 @@ export function getObjectOrJSON(
         }
       } catch (e) {
         logger.error(
-          `Encountered error:\n\n${openapiOptions}\n\nis neither a file with a valid JSON object neither a stringified JSON object.`
+          `Encountered error:\n\n${openapiOptions}\n\nis neither a file with a valid JSON object neither a stringified JSON object.\n`
         );
-        exitWithError(e);
+        throw new HandledError(e instanceof Error ? e.message : String(e));
       }
-      break;
     default: {
       if (config?.configPath) {
         logger.info(`Found ${config.configPath} and using 'openapi' options\n`);
@@ -65,7 +63,6 @@ export function getObjectOrJSON(
       return {};
     }
   }
-  return {};
 }
 
 export async function getPageHTML(